re-organized chapter 3 units
