--- src/main/kotlin/com/jeremy/kotlincourses/module3/unit3/mathops/addition.kt
@@ -1,5 +1,0 @@
-package org.example.com.jeremy.kotlinpractice.module3.unit3.mathops
-
-fun add(a: Int, b: Int): Int {
-    return a + b
-}
