--- src/main/kotlin/com/jeremy/kotlincourses/module3/unit3/stringOps/join.kt
@@ -1,5 +1,0 @@
-package org.example.com.jeremy.kotlinpractice.module3.unit3.stringOps
-
-fun concat(a: String, b: String): String {
-    return a + b
-}
