+++ src/main/kotlin/jeremy/kotlincourses/chapter2/Unit1to5.kt
@@ -1,4 +1,4 @@
-package org.example.jeremy.kotlincourses.chapter2
+package jeremy.kotlincourses.chapter2
 
 fun  main() {
     // Unit 1. Functions and Their Parameters
