+++ src/main/kotlin/jeremy/kotlincourses/chapter3/unit2/app.kt
@@ -1,7 +1,7 @@
-package org.example.com.jeremy.kotlincourses.module3.library
+package jeremy.kotlincourses.chapter3.unit2
 
 fun main(){
-    val book1: Book = Book("Hello Book","Jeremy")
+    val book1: Book = Book("Hello Book", "Jeremy")
     val book2: Book = Book("Brod", "Joe")
 //    println(book1.description)
 //    println(book2.description)
