+++ src/main/kotlin/jeremy/kotlincourses/chapter3/unit3/mathops/addition.kt
@@ -1,0 +1,5 @@
+package jeremy.kotlincourses.chapter3.unit3.mathops
+
+fun add(a: Int, b: Int): Int {
+    return a + b
+}
