+++ src/main/kotlin/jeremy/kotlincourses/chapter3/unit3/stringOps/join.kt
@@ -1,0 +1,5 @@
+package jeremy.kotlincourses.chapter3.unit3.stringOps
+
+fun concat(a: String, b: String): String {
+    return a + b
+}
