+++ src/main/kotlin/jeremy/kotlincourses/chapter3/unit5/task3/Recipe.kt
@@ -1,4 +1,4 @@
-package org.example.com.jeremy.kotlinpractice.module3.unit5.task3
+package jeremy.kotlincourses.chapter3.unit5.task3
 
 class Recipe (private val ingredients: List<String>, val name: String) {
     fun printIngredients() {
